openescape/requirements.txt
pyserial===3.2.1
ruamel===0.15.37